Change nix-shell to seperate package resolution with overrides
This will make cabal use nix-cache for depenencies,
I was not sure if that was happening before.
Overrides are easier done with hpOld/hpNew

Cabal isn't supposed to build libraries,
it should only build the project you're working on.
The nix cache is used for all dependencies.

Furthermore this should allow us to create a ci.nix
file which does all the steps and then only thing gh
actions will do is call nix-build ci.nix.
The main advantage of this is that you now can
run (and debug) CI locally.